requirements.txt
arrow==0.15.2
cefpython3==66.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pika==1.1.0
pycparser==2.19
pycryptodome==3.9.0
PyExecJS==1.5.1
pymongo==3.9.0
python-dateutil==2.8.0
redis==3.3.8
requests==2.22.0
six==1.12.0
urllib3==1.25.5
Werkzeug==0.16.0
